requirements.txt
anthropic>=0.30.0
websocket-client>=1.7.0
requests>=2.31.0
pyyaml>=6.0
